+++ translations/Aqua-mixtura_de.ts
@@ -435,17 +435,22 @@
 <context>
     <name>MixWaterWidget</name>
     <message>
-        <location filename="../src/mixture/mixwaterwidget.cpp" line="24"/>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="26"/>
         <source>Wasser</source>
         <translation>Wasser</translation>
     </message>
     <message>
-        <location filename="../src/mixture/mixwaterwidget.cpp" line="36"/>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="27"/>
+        <source>Menge:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="50"/>
         <source>Wasser hinzufügen</source>
         <translation>Wasser hinzufügen</translation>
     </message>
     <message>
-        <location filename="../src/mixture/mixwaterwidget.cpp" line="36"/>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="50"/>
         <source>Wasser löschen</source>
         <translation>Wasser löschen</translation>
     </message>
@@ -478,8 +483,8 @@
         <translation>Aufbereitung &quot;%1&quot; hat ungespeicherte Änderungen</translation>
     </message>
     <message>
+        <location filename="../src/mixture/mixturewindow.cpp" line="121"/>
         <location filename="../src/mixture/mixturewindow.cpp" line="124"/>
-        <location filename="../src/mixture/mixturewindow.cpp" line="128"/>
         <source>Aufbereitung</source>
         <translation>Aufbereitung</translation>
     </message>
@@ -833,18 +838,18 @@
 <context>
     <name>WaterSources</name>
     <message>
-        <location filename="../src/water/watersources.cpp" line="135"/>
-        <location filename="../src/water/watersources.cpp" line="143"/>
+        <location filename="../src/water/watersources.cpp" line="161"/>
+        <location filename="../src/water/watersources.cpp" line="169"/>
         <source>Name</source>
         <translation>Name</translation>
     </message>
     <message>
-        <location filename="../src/water/watersources.cpp" line="137"/>
+        <location filename="../src/water/watersources.cpp" line="163"/>
         <source>Restalkalität</source>
         <translation>Restalkalität</translation>
     </message>
     <message>
-        <location filename="../src/water/watersources.cpp" line="139"/>
+        <location filename="../src/water/watersources.cpp" line="165"/>
         <source>Menge</source>
         <translation>Menge</translation>
     </message>