+++ translations/Aqua-mixtura_en.ts
@@ -428,17 +428,22 @@
 <context>
     <name>MixWaterWidget</name>
     <message>
-        <location filename="../src/mixture/mixwaterwidget.cpp" line="24"/>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="26"/>
         <source>Wasser</source>
         <translation>Water</translation>
     </message>
     <message>
-        <location filename="../src/mixture/mixwaterwidget.cpp" line="36"/>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="27"/>
+        <source>Menge:</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="50"/>
         <source>Wasser hinzufügen</source>
         <translation>Add water</translation>
     </message>
     <message>
-        <location filename="../src/mixture/mixwaterwidget.cpp" line="36"/>
+        <location filename="../src/mixture/mixwaterwidget.cpp" line="50"/>
         <source>Wasser löschen</source>
         <translation>Delete water</translation>
     </message>
@@ -471,8 +476,8 @@
         <translation>Mixture &quot;%1&quot; has unsaved changes</translation>
     </message>
     <message>
+        <location filename="../src/mixture/mixturewindow.cpp" line="121"/>
         <location filename="../src/mixture/mixturewindow.cpp" line="124"/>
-        <location filename="../src/mixture/mixturewindow.cpp" line="128"/>
         <source>Aufbereitung</source>
         <translation>Mixture</translation>
     </message>
@@ -826,18 +831,18 @@
 <context>
     <name>WaterSources</name>
     <message>
-        <location filename="../src/water/watersources.cpp" line="135"/>
-        <location filename="../src/water/watersources.cpp" line="143"/>
+        <location filename="../src/water/watersources.cpp" line="161"/>
+        <location filename="../src/water/watersources.cpp" line="169"/>
         <source>Name</source>
         <translation>Name</translation>
     </message>
     <message>
-        <location filename="../src/water/watersources.cpp" line="137"/>
+        <location filename="../src/water/watersources.cpp" line="163"/>
         <source>Restalkalität</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../src/water/watersources.cpp" line="139"/>
+        <location filename="../src/water/watersources.cpp" line="165"/>
         <source>Menge</source>
         <translation>Amount</translation>
     </message>